+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
 * Increase default `FOLLOWED_TAGS` weight to 4
+* Add Toot.getScore() method
 
 ### v0.32.1
 * Separate out the backfill functionality to `TheAlgorithm.triggerHomeTimelineBackFill()`